# apps/control-api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImpactResponse": {
        "properties": {
          "allow_to_approval": {
            "title": "Allow To Approval",
            "type": "integer"
          },
          "allow_to_block": {
            "title": "Allow To Block",
            "type": "integer"
          },
          "approval_to_allow": {
            "title": "Approval To Allow",
            "type": "integer"
          },
          "approval_to_block": {
            "title": "Approval To Block",
            "type": "integer"
          },
          "block_to_allow": {
            "title": "Block To Allow",
            "type": "integer"
          },
          "bundle_id": {
            "title": "Bundle Id",
            "type": "string"
          },
          "content_digest": {
            "title": "Content Digest",
            "type": "string"
          },
          "evaluated_decisions": {
            "title": "Evaluated Decisions",
            "type": "integer"
          },
          "other_changes": {
            "title": "Other Changes",
            "type": "integer"
          },
          "sample_changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sample Changes",
            "type": "array"
          },
          "unchanged": {
            "title": "Unchanged",
            "type": "integer"
          }
        },
        "required": [
          "bundle_id",
          "content_digest",
          "evaluated_decisions",
          "unchanged",
          "allow_to_block",
          "allow_to_approval",
          "block_to_allow",
          "approval_to_allow",
          "approval_to_block",
          "other_changes"
        ],
        "title": "ImpactResponse",
        "type": "object"
      },
...
      "PolicyBundleInfo": {
        "properties": {
          "bundle_id": {
            "title": "Bundle Id",
            "type": "string"
          },
          "content_digest": {
            "title": "Content Digest",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "git_revision": {
            "title": "Git Revision",
            "type": "string"
          },
          "loaded_at": {
            "title": "Loaded At",
            "type": "string"
          },
          "role": {
            "default": "candidate",
            "title": "Role",
            "type": "string"
          },
          "validation_status": {
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "bundle_id",
          "content_digest",
          "git_revision",
          "loaded_at",
          "validation_status"
        ],
        "title": "PolicyBundleInfo",
        "type": "object"
      },
...
      "PolicySourceRequest": {
        "properties": {
          "digest": {
            "default": "",
            "title": "Digest",
            "type": "string"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "reference": {
            "default": "",
            "title": "Reference",
            "type": "string"
          },
          "type": {
            "default": "filesystem",
            "title": "Type",
            "type": "string"
          },
          "verify_signature": {
            "default": false,
            "title": "Verify Signature",
            "type": "boolean"
          }
        },
        "title": "PolicySourceRequest",
        "type": "object"
      },
...
      "ValidateResponse": {
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/PolicyBundleInfo"
          }
        },
        "required": [
          "bundle"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control API for private AI inference infrastructure.",
    "title": "AI Infrastructure Control Plane",
    "version": "1.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/governance/policy-bundles/validate": {
      "post": {
        "operationId": "validate_bundle_governance_policy_bundles_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicySourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Bundle",
        "tags": [
          "policy-bundles"
        ]
      }
    },
    "/governance/policy-bundles/{bundle_id}/activate": {
      "post": {
        "operationId": "activate_bundle_governance_policy_bundles__bundle_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyBundleInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Bundle",
        "tags": [
          "policy-bundles"
        ]
      }
    },
    "/governance/policy-bundles/{bundle_id}/impact": {
      "get": {
        "operationId": "get_impact_governance_policy_bundles__bundle_id__impact_get",
        "parameters": [
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Impact",
        "tags": [
          "policy-bundles"
        ]
      }
    },
    "/governance/policy-bundles/{bundle_id}/simulate": {
      "post": {
        "operationId": "simulate_bundle_governance_policy_bundles__bundle_id__simulate_post",
        "parameters": [
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Bundle",
        "tags": [
          "policy-bundles"
        ]
      }
    },
...
  }
}